% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsnes.R
\docType{package}
\name{ggsnes}
\alias{ggsnes}
\title{\code{ggsnes} package}
\description{
A collection of 'ggplot2' color palettes and themes inspired by the Super Nintendo gaming console and several popular games for the Super Nintendo.
}
\details{
See the README on
\href{https://github.com/ha0ye/ggsnes}{GitHub}
}
